+++ core/src/main/java/de/tub/dima/scotty/core/windowType/Window.java
@@ -1,7 +1,5 @@
 package de.tub.dima.scotty.core.windowType;
 
-import de.tub.dima.scotty.core.*;
-
 import java.io.*;
 
 public interface Window extends Serializable {
@@ -9,7 +7,7 @@ public interface Window extends Serializable {
 
     WindowMeasure getWindowMeasure();
 
-    default int getWindowId() {
+    default long getWindowId() {
         return NO_ID;
     }
 }
